backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic-settings==2.6.1
python-multipart==0.0.20
httpx==0.27.2
openai==1.57.4
python-dotenv==1.0.1
reka-api>=2.0.0
twilio>=9.0.0
sendgrid>=6.11.0
